@@ -1,16 +1,15 @@
-import supertest = require("supertest");
 import path = require("path");
 import { HttpStatus } from "@nestjs/common";
 
-import { connect, drop, addCategoryAndRegexp, newUser } from "../helpers/database";
-import { uploadFile } from "../helpers/files";
+import { connect, drop } from "../helpers/database";
+import * as files from "../helpers/files";
 import { sleep, newIds } from "../helpers/utils";
-import { init, initWithAuth } from "../helpers/server";
-import auth = require("@db/auth");
+import { init } from "../helpers/server";
+import { AdminRequest, GuestRequest } from "../helpers/request";
 
 describe("Files E2E Api", () => {
 
-    let request: supertest.SuperTest<supertest.Test>;
+    let request: AdminRequest;
 
     before(async () => {
         connect();
@@ -22,43 +21,42 @@ describe("Files E2E Api", () => {
         return drop(ids);
     });
 
-    before(async () => {
-        request = await init();
+    before("login", async () => {
+        const req = new GuestRequest(await init(), ids, filepaths);
+        request = await req.login();
+    });
+
+    const filepaths = [ ];
+
+    after(() => {
+        return files.remove(filepaths);
     });
 
     const getFileUrl = (cid: string, id: string) => {
         return `/files/categories/${cid}/goods/${id}`;
     };
 
-    before("login", async () => {
-        ids.users.push((await auth.login(request))[0]);
-    });
-
     let cid = "";
     let id = "";
     const uploadFilepath = `${__dirname}/../files/icon_pandorabox_64x64.png`;
     step("Upload File", async () => {
-        const docs = await addCategoryAndRegexp(/^icon_.+64x64\.png$/);
-        cid = docs[0]._id;
-        const { body: result, status } =
-            await uploadFile(request, uploadFilepath);
+        const docs = await request.addCategoryWithRegexp(/^icon_.+64x64\.png$/);
+        cid = ids.categories[ids.categories.length - 1].toString();
+        const { status } = await request.uploadFile(uploadFilepath);
+        id = ids.goods[ids.goods.length - 1].toString();
         status.should.be.eql(201);
-        id = result._id;
-        ids.categories.push(docs[0]._id);
-        ids.regexps.push(docs[1]._id);
-        ids.goods.push(id);
         await sleep(50);
     });
 
     step("Logout", () => {
-        return request.get("/api/v1/auth/logout").then();
+        return request.logout();
     });
 
     step("Download File", async () => {
         const filename = path.basename(uploadFilepath);
 
-        const url = getFileUrl(cid, id);
-        const { status, header, body } = await request.get(url).then();
+        const { status, header, body } =
+            await request.downloadFile(cid, id);
 
         status.should.eql(HttpStatus.OK);
         header.should.match({
@@ -69,8 +67,9 @@ describe("Files E2E Api", () => {
     step("Download Nonexist File", async () => {
         const cid = "5a44d78fec77afe7c8aa3eca";
         const id = "5a44d78fec77afe7c8aa3eca";
-        const url = getFileUrl(cid, id);
-        const result = await request.get(url).redirects(1).then();
+        const result = await request.downloadFile(cid, id, {
+            redirects: 1
+        });
 
         result.should.have.property("status", HttpStatus.NOT_FOUND);
     });
@@ -78,8 +77,9 @@ describe("Files E2E Api", () => {
     step("Download Wrong ID File #0", async () => {
         const cid = "5a44d78fec77afe7c8aa3eca";
         const id = "1111";
-        const url = getFileUrl(cid, id);
-        const result = await request.get(url).redirects(1).then();
+        const result = await request.downloadFile(cid, id, {
+            redirects: 1
+        });
 
         result.should.have.property("status", HttpStatus.BAD_REQUEST);
     });
@@ -87,8 +87,9 @@ describe("Files E2E Api", () => {
     step("Download Wrong ID File #1", async () => {
         const cid = "1111";
         const id = "5a44d78fec77afe7c8aa3eca";
-        const url = getFileUrl(cid, id);
-        const result = await request.get(url).redirects(1).then();
+        const result = await request.downloadFile(cid, id, {
+            redirects: 1
+        });
 
         result.should.have.property("status", HttpStatus.BAD_REQUEST);
     });
