@@ -1,16 +1,13 @@
-import supertest = require("supertest");
-
-import { connect, drop, newUser } from "../helpers/database";
+import { connect, drop } from "../helpers/database";
 import { init } from "../helpers/server";
 import { UsersService } from "@services/users";
-import { TokensService } from "@services/tokens";
 import { SystemService } from "@services/system";
-import { newName, newIds } from "../helpers/utils";
+import { newIds } from "../helpers/utils";
+import { TokenRequest, GuestRequest } from "../helpers/request";
 
 describe("Fix Issues", () => {
 
-    let request: supertest.SuperTest<supertest.Test>;
-    const tokensSvr = new TokensService();
+    let request: TokenRequest;
     const usersSvr = new UsersService(new SystemService());
 
     before(() => {
@@ -23,34 +20,14 @@ describe("Fix Issues", () => {
         return drop(ids);
     });
 
-    before(async () => {
-        request = await init();
+    before("Login", async () => {
+        request = await new GuestRequest(await init(), ids).loginWithToken();
     });
 
-    const user = {
-        name: newName(),
-        pass: newName(),
-        token: ""
-    };
     describe("Token Action When User ban", () => {
 
-        step("Login", async () => {
-            const doc = await newUser(user.name, user.pass);
-            ids.users.push(doc._id);
-            const {
-                body: result
-            } = await request.post("/api/v1/auth/login?token=true")
-                .send({
-                    username: user.name, password: user.pass
-                }).then();
-            result.should.have.property("token");
-            ids.tokens.push(await tokensSvr.getIdByToken(result.token));
-            user.token = result.token;
-        });
-
         step("Get Goods Success By Token", async () => {
-            const { status } = await request.get("/api/v1/users/goods")
-            .auth(user.name, user.token).then();
+            const { status } = await request.get("/api/v1/users/goods").then();
             status.should.be.eql(200);
         });
 
@@ -59,8 +36,7 @@ describe("Fix Issues", () => {
         });
 
         step("Get Goods Fail By Token", async () => {
-            const { status } = await request.get("/api/v1/users/goods")
-                .auth(user.name, user.token).then();
+            const { status } = await request.get("/api/v1/users/goods").then();
             status.should.be.eql(403);
         });
 
@@ -69,8 +45,7 @@ describe("Fix Issues", () => {
         });
 
         step("Get Goods Fail By Token", async () => {
-            const { status } = await request.get("/api/v1/users/goods")
-                .auth(user.name, user.token).then();
+            const { status } = await request.get("/api/v1/users/goods").then();
             status.should.be.eql(200);
         });
 
