@@ -2,7 +2,7 @@ import { Module, MiddlewaresConsumer } from "@nestjs/common";
 // Modules
 import { DatabaseModule } from "./database/database.module";
 import {controllers, ControllersModule} from "./controllers.module";
-import { NoCacheMiddleware } from "./common/middlewares/noCache.middleware";
+import { NoCacheMiddleware } from "../middlewares/noCache.middleware";
 
 @Module({
     modules: [ DatabaseModule, ControllersModule ]
