@@ -21,14 +21,14 @@ import { SystemController } from "./system/system.controller";
 // region Middlewares
 import {
     UploadFileMiddleware, UploadFilesMiddleware
-} from "./common/middlewares/upload.middleware";
+} from "../middlewares/upload.middleware";
 import {
     ApiLoggerMiddleware, DownloadLoggerMiddleware
-} from "./common/middlewares/logger.middleware";
-import { RolesMiddleware } from "./common/middlewares/roles.middleware";
+} from "../middlewares/logger.middleware";
+import { RolesMiddleware } from "../middlewares/roles.middleware";
 import {
     ReloadSessionMiddleware
-} from "./common/middlewares/reloadSession.middleware";
+} from "../middlewares/reloadSession.middleware";
 // endregion Middlewares
 
 // region Services
@@ -40,6 +40,7 @@ import { UsergroupsService } from "@services/usergroups";
 import { SystemService } from "@services/system";
 import { CategoriesService } from "@services/categories";
 import { GoodsService } from "@services/goods";
+import { LogsService } from "@services/logs";
 // endregion Services
 
 export const controllers = [
@@ -56,7 +57,7 @@ export const controllers = [
 export const services = [
     RegexpsService, CategoriesService, GoodsService,
     CollectionsService, TokensService, UsersService, UsergroupsService,
-    SystemService
+    SystemService, LogsService
 ];
 
 @Module({
