@@ -34,6 +34,7 @@ import { isArray } from "util";
 import { CreateValueDto, EditValueDto } from "../values/values.dto";
 import { GoodAttributeParamDto, UploadQueryDto, EditBodyDto } from "./goods.dto";
 import { RegexpCountCheckInterceptor } from "@interceptors/regexp-count-check";
+import { LogsService } from "@services/logs";
 
 @UseGuards(RolesGuard)
 @Controller("api/v1/goods")
@@ -47,7 +48,8 @@ export class GoodsAdminController {
         private readonly collectionsSvr: CollectionsService,
         private readonly regexpSvr: RegexpsService,
         private readonly categoriesSvr: CategoriesService,
-        private readonly goodsSvr: GoodsService
+        private readonly goodsSvr: GoodsService,
+        private readonly logsSvr: LogsService
     ) { }
 
     private toMd5sum(filepath: string) {
@@ -132,6 +134,7 @@ export class GoodsAdminController {
                 active: true
             })).toObject();
         } catch (error) {
+            fs.remove(obj.file.path);
             if (cb) {
                 cb("Good", error);
             } else {
@@ -168,7 +171,7 @@ export class GoodsAdminController {
             appends: await this.getCategoriesIds(query.append || [])
         };
 
-        return await this.fileProcess(
+        return this.fileProcess(
             { file, uploader: uploaderId, opt: fileProcessOpts },
             (type, error) => {
                 if (type === "Categories") {
@@ -179,7 +182,7 @@ export class GoodsAdminController {
                     }
                 }
                 if (type === "Good") {
-                    throw new BadRequestException(error.toString());
+                    throw error;
                 }
             }
         );
@@ -232,7 +235,7 @@ export class GoodsAdminController {
             });
             const collection =
                 isArray(collections) ? collections[0] : collections;
-            return this.collectionsSvr.getById(collection._id, {
+            return this.collectionsSvr.getObjectById(collection._id, {
                 populate: [ "goods" ]
             });
         }
@@ -247,13 +250,14 @@ export class GoodsAdminController {
     public async get(@Param() param: GidDto) {
         let obj;
         try {
-            obj = await this.goodsSvr.getById(param.gid, {
+            obj = await this.goodsSvr.getObjectById(param.gid, {
                 populate: [
                     "attributes",
                     { path: "uploader", select: "nickname" },
                     { path: "category", select: "name attributes tags" }
                 ]
             });
+            obj.downloaded = await this.logsSvr.goodDownloadCount(param.gid);
         } catch (error) {
             throw new BadRequestException(error.toString());
         }
