@@ -1,7 +1,7 @@
 import { Component, Param, BadRequestException } from "@nestjs/common";
 import { UidDto } from "@dtos/ids";
 import {
-    Model as CollectionsModel, cache, ICollections
+    Model as CollectionsModel, cache, ICollections, ICollectionsRaw
 } from "@models/Collection";
 import { ObjectId } from "@models/common";
 import { DEF_PER_COUNT } from "@dtos/page";
@@ -51,10 +51,10 @@ export class CollectionsService extends BaseService<ICollections> {
      * Get By Collection Name
      * @param name Collection Name
      */
-    public getByName(name: string, opts = this.GET_OPTIONS) {
+    public getObjectByName(name: string, opts = this.GET_OPTIONS) {
         return this.loadAndCache(
             `getByName_${name}`,
-            () => this.findObject({ name }, opts),
+            () => this.findObject({ name }, opts) as Promise<ICollectionsRaw>,
             1000
         );
     }
@@ -63,10 +63,10 @@ export class CollectionsService extends BaseService<ICollections> {
      * Get By Collection ID
      * @param id Collection ID
      */
-    public getById(id: ObjectId, opts = this.GET_OPTIONS) {
+    public getObjectById(id: ObjectId, opts = this.GET_OPTIONS) {
         return this.loadAndCache(
             `getById_${id.toString()}`,
-            () => this.findObjectById(id, opts),
+            () => this.findObjectById(id, opts) as Promise<ICollectionsRaw>,
             1000
         );
     }
