@@ -1,63 +1,50 @@
-import supertest = require("supertest");
 import fs = require("fs-extra");
 import { basename } from "path";
 
-import { Model as GoodsModels } from "@models/Good";
 import {
-    connect, drop, newUser, addCategoryAndRegexp
+    connect, drop, addCategoryAndRegexp
 } from "../../helpers/database";
 import { init } from "../../helpers/server";
 import * as files from "../../helpers/files";
 
-import { config } from "@utils/config";
-import auth = require("@db/auth");
 import { newIds } from "../../helpers/utils";
+import { AdminRequest, GuestRequest } from "../../helpers/request";
 
 /**
  * Fix [Issue 31](https://github.com/BoxSystem/StoreBox-Api/issues/31)
  */
 describe("Fix Issues", () => {
 
-    let request: supertest.SuperTest<supertest.Test>;
+    let request: AdminRequest;
 
     before(() => {
         return connect();
     });
 
     const ids = newIds();
+    const filepaths = [ ];
 
     after(() => {
         return drop(ids);
     });
 
+    before("login", async () => {
+        request = await new GuestRequest(await init(), ids, filepaths).login();
+    });
+
     before(async () => {
-        request = await init();
+        await request.newFile();
+    });
+
+    after(() => {
+        return files.remove(filepaths);
     });
 
     describe("Github 31 [Can upload same file]", () => {
 
-        let filepath = "";
-        let filename = "";
-        before(async () => {
-            filepath = await files.newFile();
-        });
-
-        after(() => {
-            return files.remove(filepath);
-        });
-
-        after(() => {
-            GoodsModels.remove({
-                originname: filename
-            }).exec();
-        });
-
-        before("login", async () => {
-            ids.users.push((await auth.login(request))[0]);
-        });
-
         step("Add Category and Regexp", async () => {
-            filename = basename(filepath);
+            const filepath = filepaths[filepaths.length - 1];
+            const filename = basename(filepath);
             const docs = await addCategoryAndRegexp(
                 new RegExp(`^${filename}$`)
             );
@@ -66,16 +53,12 @@ describe("Fix Issues", () => {
         });
 
         step("Upload Success", async () => {
-            const {
-                body: result, status
-            } = await files.uploadFile(request, filepath);
+            const { status } = await request.uploadFile();
             status.should.be.eql(201);
         });
 
         step("Upload Fail", async () => {
-            const {
-                body: result, status
-            } = await files.uploadFile(request, filepath);
+            const { status, body } = await request.uploadFile();
             status.should.be.not.eql(201);
             status.should.be.eql(400);
         });
